Move fn to client
